refactor: split journal archivist.ts into cli + schemas (PR3 of #799)
The 386-line archivist.ts had grown into a hotel mixing:
- the Claude CLI subprocess wrapper
- the Summarize transport type and ClaudeCliNotFoundError /
  ClaudeCliFailedError classes
- 6 large interfaces (SessionEventExcerpt, SessionExcerpt,
  ExistingTopicSnapshot, DailyArchivistInput, TopicUpdate, etc.)
- 2 long system prompts
- 2 user-prompt builders
- 2 type guards (isDailyArchivistOutput, isOptimizationOutput)
- a re-export of extractJsonObject / findBalancedBraceBlock

The split:
- `archivist-cli.ts` — Summarize type, error classes, runClaudeCli.
  The only journal file that imports node:child_process. Tests can
  inject a fake Summarize without ever touching subprocess code.
- `archivist-schemas.ts` — interfaces, prompts, prompt builders,
  validators. Pure module — no IO, no subprocess, no global state.

12 importers updated to point at whichever new file they need.
Most cross-subsystem callers (chat-index, sources) only need
ClaudeCliNotFoundError, so they import from archivist-cli now.
The journal-internal callers split their imports between the two
files based on what they use.

Old `archivist.ts` deleted — no shim. The audit (#799) explicitly
called out re-export shims as clutter to remove.

Touched files:
  server/workspace/journal/{archivist-cli,archivist-schemas}.ts (new)
  server/workspace/journal/archivist.ts (deleted)
  server/workspace/journal/{dailyPass,index,memoryExtractor,
                            optimizationPass}.ts (imports)
  server/workspace/chat-index/{index,summarizer}.ts (imports)
  server/workspace/sources/{classifier,pipeline/summarize}.ts (imports)
  test/journal/{test_archivist,test_archivist_extract,
                test_dailyPass}.ts (imports)
  test/chat-index/test_maybe_index_session.ts (imports)

Behaviour-preserving: 3084/3084 unit tests pass with no test logic
changes — only import path updates.

Refs #799.
