+++ server/workspace/chat-index/summarizer.ts
@@ -17,7 +17,7 @@ import { EVENT_TYPES } from "../../../src/types/events.js";
 import { readFile } from "node:fs/promises";
 import { formatSpawnFailure } from "../../utils/spawn.js";
 import { tmpdir } from "node:os";
-import { ClaudeCliNotFoundError } from "../journal/archivist.js";
+import { ClaudeCliNotFoundError } from "../journal/archivist-cli.js";
 import { errorMessage } from "../../utils/errors.js";
 import type { SummaryResult } from "./types.js";
 import { ONE_MINUTE_MS } from "../../utils/time.js";
